+++ crates/idxdb-store/src/account/utils.rs
@@ -17,38 +17,38 @@ use super::js_bindings::{
 };
 use super::models::AccountRecordIdxdbObject;
 
-pub async fn insert_account_code(account_code: &AccountCode) -> Result<(), JsValue> {
+pub async fn upsert_account_code(account_code: &AccountCode) -> Result<(), JsValue> {
     let root = account_code.commitment().to_string();
     let code = account_code.to_bytes();
 
-    let promise = idxdb_insert_account_code(root, code);
+    let promise = idxdb_upsert_account_code(root, code);
     JsFuture::from(promise).await?;
 
     Ok(())
 }
 
-pub async fn insert_account_storage(account_storage: &AccountStorage) -> Result<(), JsValue> {
+pub async fn upsert_account_storage(account_storage: &AccountStorage) -> Result<(), JsValue> {
     let root = account_storage.commitment().to_string();
 
     let storage = account_storage.to_bytes();
 
-    let promise = idxdb_insert_account_storage(root, storage);
+    let promise = idxdb_upsert_account_storage(root, storage);
     JsFuture::from(promise).await?;
 
     Ok(())
 }
 
-pub async fn insert_account_asset_vault(asset_vault: &AssetVault) -> Result<(), JsValue> {
+pub async fn upsert_account_asset_vault(asset_vault: &AssetVault) -> Result<(), JsValue> {
     let commitment = asset_vault.root().to_string();
     let assets = asset_vault.assets().collect::<Vec<Asset>>().to_bytes();
 
-    let promise = idxdb_insert_account_asset_vault(commitment, assets);
+    let promise = idxdb_upsert_account_asset_vault(commitment, assets);
     JsFuture::from(promise).await?;
 
     Ok(())
 }
 
-pub async fn insert_account_record(
+pub async fn upsert_account_record(
     account: &Account,
     account_seed: Option<Word>,
 ) -> Result<(), JsValue> {
@@ -61,7 +61,7 @@ pub async fn insert_account_record(
     let account_seed = account_seed.map(|seed| seed.to_bytes());
     let commitment = account.commitment().to_string();
 
-    let promise = idxdb_insert_account_record(
+    let promise = idxdb_upsert_account_record(
         account_id_str,
         code_root,
         storage_root,
@@ -107,7 +107,7 @@ pub fn parse_account_record_idxdb_object(
 }
 
 pub async fn update_account(new_account_state: &Account) -> Result<(), JsValue> {
-    insert_account_storage(new_account_state.storage()).await?;
-    insert_account_asset_vault(new_account_state.vault()).await?;
-    insert_account_record(new_account_state, None).await
+    upsert_account_storage(new_account_state.storage()).await?;
+    upsert_account_asset_vault(new_account_state.vault()).await?;
+    upsert_account_record(new_account_state, None).await
 }